# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "golang crud api",
        "title": "gocrud",
        "contact": {},
        "version": "1.0"
    },
    "host": "{host}:{port}",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Retrieve data of all registered users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get all users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter users by ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter users by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all user data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.UserPaginationResponse"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Retrieve data of a user based on the provided ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user by ID",
                        "schema": {
                            "$ref": "#/definitions/domain.UserResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update data of an existing user based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user data by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated user data object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated user by ID",
                        "schema": {
                            "$ref": "#/definitions/domain.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorValidationResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing user based on the provided ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete a user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted user by ID",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Claims": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.ErrorValidationResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ValidationError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Metadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "order": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                },
                "page": {
                    "type": "integer"
                },
                "sort": {
                    "type": "string",
                    "enum": [
                        "id",
                        "user_id",
                        "name",
                        "title",
                        "created_at",
                        "updated_at"
                    ]
                },
                "total_page": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.UserPaginationResponse": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/domain.Metadata"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserResponse"
                    }
                }
            }
        },
...
        "domain.UserRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string",
                    "maxLength": 50
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 4
                },
                "password": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 8
                }
            }
        },
        "domain.UserResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tokens": {
                    "$ref": "#/definitions/domain.UserToken"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.UserToken": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "claims": {
                    "$ref": "#/definitions/domain.Claims"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "domain.ValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        }
    }
}